+++ extensions/iceworks-time-master/src/config.ts
@@ -1,0 +1,9 @@
+import { ONE_MIN_MILLISECONDS } from './constants';
+
+export const jsonSpaces = 2;
+export const recordKeystrokeDuration = 1;
+export const recordKeystrokeDurationMins = ONE_MIN_MILLISECONDS * recordKeystrokeDuration;
+export const checkMidnightDuration = 5;
+export const checkMidnightDurationMins = ONE_MIN_MILLISECONDS * checkMidnightDuration;
+export const sendPayloadDuration = 15;
+export const snedPayloadDurationMins = ONE_MIN_MILLISECONDS * sendPayloadDuration;