+++ extensions/iceworks-time-master/src/recorders/keystrokeStats/index.ts
@@ -1,0 +1,2 @@
+export { KeystrokeStatsRecorder } from './recorder';
+export { KeystrokeStats, KeystrokeStatsInfo } from './keystrokeStats';