+++ extensions/iceworks-time-master/src/storages/average.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 import * as fse from 'fs-extra';
-import { getAppDataDirPath, getStorageDirs } from '../utils/storage';
+import { getStoragePath, getStorageDaysDirs } from '../utils/storage';
 import { getCountAndAverage4UserSummary } from './user';
-import { JSON_SPACES } from '../constants';
+import { jsonSpaces } from '../config';
 
 export class AverageSummary {
   dailySessionSeconds?: number = 0;
@@ -15,7 +15,7 @@ export class AverageSummary {
 }
 
 export function getAverageFile() {
-  return path.join(getAppDataDirPath(), 'average.json');
+  return path.join(getStoragePath(), 'average.json');
 }
 
 export async function getAverageSummary(): Promise<AverageSummary> {
@@ -31,7 +31,7 @@ export async function getAverageSummary(): Promise<AverageSummary> {
 
 export async function saveAverageSummary(averageSummary: AverageSummary) {
   const file = getAverageFile();
-  await fse.writeJson(file, averageSummary, { spaces: JSON_SPACES });
+  await fse.writeJson(file, averageSummary, { spaces: jsonSpaces });
 }
 
 export async function clearAverageSummary() {
@@ -40,7 +40,7 @@ export async function clearAverageSummary() {
 }
 
 export async function updateAverageSummary(): Promise<AverageSummary> {
-  const storageDirs = await getStorageDirs();
+  const storageDirs = await getStorageDaysDirs();
 
   // Don't add today to the average
   if (storageDirs.length > 0) {