+++ extensions/iceworks-time-master/src/types.d.ts
@@ -1,1 +1,3 @@
 declare module '@iceworks/common-service';
+
+type PropType<Obj, Prop extends keyof Obj> = Obj[Prop];