+++ packages/constant/src/index.ts
@@ -23,3 +23,4 @@ export const ALI_TASKRESULT_URL = Base64.decode(
 export const ALI_APPLYREPO_URL = Base64.decode('aHR0cHM6Ly9hcGkuZGVmLmFsaWJhYmEtaW5jLmNvbS9hcGkvd29yay9yZXBvL2FwcGx5');
 export const ALI_GITLAB_URL = Base64.decode('Z2l0QGdpdGxhYi5hbGliYWJhLWluYy5jb20=');
 export const ALI_EMAIL = Base64.decode('YWxpYmFiYS1pbmMuY29t');
+export const ALI_DIP_DAILY = Base64.decode('aHR0cDovL2ljZXdvcmtzLWlkcC5meC5hbGliYWJhLm5ldA==');