+++ extensions/iceworks-time-master/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.2.0
+
+- feat: refresh views when window is focused
+- feat: check is sendable
+
 ## 0.1.0
 
 - Initial release